+++ serial_gca.h
@@ -2,5 +2,5 @@
 #define _SERIAL_GCA_H
 #include <stdbool.h>
 
-float* serial_gca(float (*target_function)(float*, int), float lb, float ub, int dim, int pop_size, int n_iter, bool debug);
+double* serial_gca(double (*target_function)(double*, int), double lb, double ub, int dim, int pop_size, int n_iter, bool debug);
 #endif
