+++ test_functions.c
@@ -1,6 +1,6 @@
 
-float sphere(float* x, int dim) {
-    float sum = 0;
+double sphere(double* x, int dim) {
+    double sum = 0;
     for (int i = 0; i < dim; i++) {
         sum += x[i] * x[i];
     }
