+++ test_functions.h
@@ -1,5 +1,5 @@
 #ifndef _TEST_FUNCTIONS_H
 #define _TEST_FUNCTIONS_H
 
-float sphere(float* x, int dim);
+double sphere(double* x, int dim);
 #endif
